## Add

- **Message fragmentation**: enable sending/receiving messages larger than the MTU setup in the transport splitting these into fragments.
- **Time synchronization**: allows synchronizing the Client with the Agent through the NTP protocol.
- **Windows discovery**: allows discovering Agents using multicast/unicast in Windows systems.
- **New unitary tests**: check the correct functioning of Client's core.
- **API documentation**: as comments in the code in order to export to Doxygen page.
- **Raspberry Pi support**: allows compiling the Client in Raspbian systems.

## Change
- **Memory usage**: is optimized for Linux platform using the `-fdata-sections` and `-ffunction-sections`.
- **CMake approach**: now used the superbuild pattern instead of add_subdirectory to solve the Micro CDR dependency.
- **Discovery API**: is simplified removing the `timestamp` parameter in the callback function and the `chosen` parameter in the discovery functions.
- **Examples usage**: now accept the IP and port as arguments being more user-friendly.

## Fix
* **Acknack reading**: was modified fixing the reliable communication state-machine.

+++ examples/Deployment/HelloWorld.c
@@ -45,9 +45,9 @@ bool HelloWorld_deserialize_topic(ucdrBuffer* reader, HelloWorld* topic)
 uint32_t HelloWorld_size_of_topic(const HelloWorld* topic, uint32_t size)
 {
     uint32_t previousSize = size;
-    size += ucdr_alignment(size, 4) + 4;
+    size += (uint32_t)(ucdr_alignment(size, 4) + 4);
 
-    size += ucdr_alignment(size, 4) + 4 + (uint32_t)strlen(topic->message) + 1;
+    size += (uint32_t)(ucdr_alignment(size, 4) + 4 + strlen(topic->message) + 1);
 
     return size - previousSize;
 }
