+++ examples/ShapesDemo/ShapeType.c
@@ -53,13 +53,13 @@ bool ShapeType_deserialize_topic(ucdrBuffer* reader, ShapeType* topic)
 uint32_t ShapeType_size_of_topic(const ShapeType* topic, uint32_t size)
 {
     uint32_t previousSize = size;
-    size += ucdr_alignment(size, 4) + 4 + (uint32_t)strlen(topic->color) + 1;
+    size += (uint32_t)(ucdr_alignment(size, 4) + 4 + strlen(topic->color) + 1);
 
-    size += ucdr_alignment(size, 4) + 4;
+    size += (uint32_t)(ucdr_alignment(size, 4) + 4);
 
-    size += ucdr_alignment(size, 4) + 4;
+    size += (uint32_t)(ucdr_alignment(size, 4) + 4);
 
-    size += ucdr_alignment(size, 4) + 4;
+    size += (uint32_t)(ucdr_alignment(size, 4) + 4);
 
     return size - previousSize;
 }
